# apps/ims-server/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserResponseType": {
				"properties": {
					"identifierType": {
						"type": "string",
						"enum": [
							"email",
							"phone"
						]
					},
					"identifier": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"identifierType",
					"identifier",
					"name",
					"id"
				],
				"type": "object"
			},
			"RefreshTokenInput": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"_36_Enums.OtpType": {
				"type": "string",
				"enum": [
					"SIGN_UP",
					"LOGIN",
					"FORGET_PASSWORD"
				]
			},
			"OtpType": {
				"$ref": "#/components/schemas/_36_Enums.OtpType"
			},
			"AuthTokensResponseType": {
				"properties": {
					"refreshToken": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken",
					"accessToken"
				],
				"type": "object"
			},
			"AuthResponseType": {
				"properties": {
					"tokens": {
						"$ref": "#/components/schemas/AuthTokensResponseType"
					},
					"user": {
						"$ref": "#/components/schemas/UserResponseType"
					}
				},
				"required": [
					"tokens",
					"user"
				],
				"type": "object"
			},
			"LocalLoginInput": {
				"properties": {
					"password": {
						"type": "string"
					},
					"identifier": {
						"type": "string"
					}
				},
				"required": [
					"password",
					"identifier"
				],
				"type": "object"
			},
			"OtpSignUpInput": {
				"properties": {
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"password"
				],
				"type": "object"
			},
			"ResetForgettenPasswordInput": {
				"properties": {
					"confirmNewPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"confirmNewPassword",
					"newPassword"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@khaled/ims-server",
		"version": "1.3.0",
		"contact": {}
	},
	"paths": {
		"/user": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponseType"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/oauth/google": {
			"get": {
				"operationId": "GoogleLogin",
				"responses": {
					"302": {
						"description": "Redirect"
					}
				},
				"tags": [
					"oauth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/oauth/facebook": {
			"get": {
				"operationId": "FacebookLogin",
				"responses": {
					"302": {
						"description": "Redirect"
					}
				},
				"tags": [
					"oauth"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/token/refresh": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"accessToken": {
											"type": "string"
										}
									},
									"required": [
										"accessToken"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"token"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RefreshTokenInput"
							}
						}
					}
				}
			}
		},
		"/otp/request": {
			"post": {
				"operationId": "RequestOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otpType": {
										"$ref": "#/components/schemas/OtpType"
									},
									"identifier": {
										"type": "string"
									}
								},
								"required": [
									"otpType",
									"identifier"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/otp/verify": {
			"post": {
				"operationId": "VerifyOtp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										}
									},
									"required": [
										"token"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"otpType": {
										"$ref": "#/components/schemas/OtpType"
									},
									"otp": {
										"type": "string"
									}
								},
								"required": [
									"otpType",
									"otp"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponseType"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LocalLoginInput"
							}
						}
					}
				}
			}
		},
		"/auth/sign-up": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthResponseType"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OtpSignUpInput"
							}
						}
					}
				}
			}
		},
		"/auth/forget-password": {
			"post": {
				"operationId": "ResetForgettenPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetForgettenPasswordInput"
							}
						}
					}
				}
			}
		},
		"/auth/logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"refreshToken": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/health-check": {
			"get": {
				"operationId": "HealthCheck",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"description": "Simple health check endpoint",
				"tags": [
					"Health"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}