# apps/ims-web/src/api/model/authResponseType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * @khaled/ims-server
 * OpenAPI spec version: 1.3.0
 */
import type { AuthTokensResponseType } from './authTokensResponseType';
import type { UserResponseType } from './userResponseType';

export interface AuthResponseType {
  tokens: AuthTokensResponseType;
  user: UserResponseType;
}