# apps/ims-web/src/api/model/authTokensResponseType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * @khaled/ims-server
 * OpenAPI spec version: 1.3.0
 */

export interface AuthTokensResponseType {
  refreshToken: string;
  accessToken: string;
}