# apps/ims-web/src/api/model/localLoginInput.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * @khaled/ims-server
 * OpenAPI spec version: 1.3.0
 */

export interface LocalLoginInput {
  password: string;
  identifier: string;
}