# apps/ims-web/src/api/model/otpType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * @khaled/ims-server
 * OpenAPI spec version: 1.3.0
 */
import type { _36EnumsOtpType } from './_36EnumsOtpType';

export type OtpType = _36EnumsOtpType;