# apps/ims-web/src/api/model/userResponseType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * @khaled/ims-server
 * OpenAPI spec version: 1.3.0
 */
import type { UserResponseTypeIdentifierType } from './userResponseTypeIdentifierType';

export interface UserResponseType {
  identifierType: UserResponseTypeIdentifierType;
  identifier: string;
  name: string;
  id: string;
}